Dissociates dispatching and closure capture identification
